# openapi.yaml
openapi: 3.1.0
info:
  title: Memoria API
  version: 1.0.0
  description: Private photo/video sharing API (Capsules + Albums). All times UTC in JSON.
servers:
  - url: http://localhost:8080
    description: Local docker-compose
  - url: https://api.memoria.app
    description: Production (placeholder)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
    TokenPair:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        expires_in: { type: integer }
...
paths:
...
  /v1/auth/otp/request:
    post:
      tags: [auth]
      summary: Request OTP email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, purpose]
              properties:
                email: { type: string, format: email }
                purpose: { type: string, enum: [signup, password_reset] }
      responses:
        '200': { description: Code sent }
        '429': { description: Rate limited, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
  /v1/auth/otp/verify:
    post:
      tags: [auth]
      summary: Verify OTP → ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, purpose, code]
              properties:
                email: { type: string }
                purpose: { type: string }
                code: { type: string }
      responses:
        '200':
          description: Ticket issued
  /v1/auth/signup:
    post:
      tags: [auth]
      summary: Create account with signup ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ticket, password, username, display_name]
              properties:
                ticket: { type: string }
                password: { type: string, minLength: 8, maxLength: 128 }
                username: { type: string }
                display_name: { type: string, maxLength: 100 }
                timezone: { type: string }
                device_name: { type: string }
      responses:
        '201': { description: User + tokens }
  /v1/auth/username-available:
    get:
      tags: [auth]
      summary: Check username availability
      parameters:
        - name: u
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: available boolean }
  /v1/auth/login:
    post:
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
                device_name: { type: string }
      responses:
        '200': { description: User + tokens }
  /v1/auth/reset-password:
    post:
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ticket, new_password]
              properties:
                ticket: { type: string }
                new_password: { type: string }
      responses:
        '200': { description: Password updated }
  /v1/auth/refresh:
    post:
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: Rotated tokens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
  /v1/auth/logout:
    post:
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200': { description: Logged out }
...